backend/requirements.txt
# FastAPI backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
python-multipart>=0.0.12

# Transcription (WhisperX for word-level alignment)
whisperx>=3.1.0
faster-whisper>=1.0.0

# Audio / Video processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0
soundfile>=0.10.3

# ML / GPU
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0

# Speaker diarization
pyannote.audio>=3.1.1

# AI providers
openai>=1.50.0
anthropic>=0.39.0
requests>=2.28.0

# Audio cleanup
deepfilternet>=0.5.0

# Utilities
pydantic>=2.0.0
